--- a/DFC.ServiceTaxonomy.Editor/NLog.dev.config
+++ b/DFC.ServiceTaxonomy.Editor/NLog.dev.config
@@ -33,6 +33,7 @@
         <logger name="DFC.ServiceTaxonomy.GraphSync.GraphSyncers.DeleteGraphSyncer" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.Neo4j.Log.NeoLogger" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.Neo4j.Services.NeoGraphDatabase" minlevel="Info" writeTo="file" />
+        <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.ContentNoCacheStep" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.CypherCommandStep" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.CypherToContentStep" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.Events.Services.EventGridContentClient" minlevel="Info" writeTo="file" />