--- a/DFC.ServiceTaxonomy.Editor/NLog.dev.config
+++ b/DFC.ServiceTaxonomy.Editor/NLog.dev.config
@@ -36,6 +36,7 @@
         <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.ContentNoCacheStep" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.CypherCommandStep" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.CypherToContentStep" minlevel="Info" writeTo="file" />
+        <logger name="DFC.ServiceTaxonomy.GraphSync.Recipes.Executors.CSharpContentStep" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.Events.Services.EventGridContentClient" minlevel="Info" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.GraphSync.GraphSyncers.MergeGraphSyncer" minlevel="Debug" writeTo="file" />
         <logger name="DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Parts.Flow.FlowPartEmbeddedContentItemsGraphSyncer" minlevel="Debug" writeTo="file" />