--- a/GetJobProfiles/ContentRecipes/ContactUsPages.recipe.json
+++ b/GetJobProfiles/ContentRecipes/ContactUsPages.recipe.json
@@ -10,7 +10,7 @@
     "tags": [],
     "steps": [
         {
-            "name": "content",
+            "name": "ContentNoCache",
             "data": [
                 {
                     "ContentItemId": "4yhev1xmryp1mvjkd9693kwqc9",