--- a/GetJobProfiles/ContentRecipes/EmailTemplates.recipe.json
+++ b/GetJobProfiles/ContentRecipes/EmailTemplates.recipe.json
@@ -10,7 +10,7 @@
     "tags": [],
     "steps": [
         {
-            "name": "content",
+            "name": "ContentNoCache",
             "data": [
                 {
                     "ContentItemId": "4kxntsjh5rncv5dqd67q7aaam6",