--- a/GetJobProfiles/ContentRecipes/SharedContent.recipe.json
+++ b/GetJobProfiles/ContentRecipes/SharedContent.recipe.json
@@ -10,7 +10,7 @@
     "tags": [],
     "steps": [
         {
-            "name": "CSharpContent",
+            "name": "ContentNoCache",
             "data": [
                 {
                     "ContentItemId": "4kqbfyyy1ahd5xndf45fg2rad0",