--- a/GetJobProfiles/ContentRecipes/SkillsToolkit.recipe.json
+++ b/GetJobProfiles/ContentRecipes/SkillsToolkit.recipe.json
@@ -10,7 +10,7 @@
     "tags": [],
     "steps": [
         {
-            "name": "CSharpContent",
+            "name": "ContentNoCache",
             "data": [
                 {
                     "ContentItemId": "4bhp0c395zz6hspv6webkhp2dj",