--- a/GetJobProfiles/ContentRecipes/Taxonomies.recipe.json
+++ b/GetJobProfiles/ContentRecipes/Taxonomies.recipe.json
@@ -10,7 +10,7 @@
     "tags": [],
     "steps": [
         {
-            "name": "CSharpContent",
+            "name": "ContentNoCache",
             "data": [
                 {
                     "ContentItemId": "4eembshqzx66drajtdten34tc8",