--- a/GetJobProfiles/ContentRecipes/TestPages.recipe.json
+++ b/GetJobProfiles/ContentRecipes/TestPages.recipe.json
@@ -10,7 +10,7 @@
     "tags": [],
     "steps": [
         {
-            "name": "CSharpContent",
+            "name": "ContentNoCache",
             "data": [
                 {
                     "ContentItemId": "4y795j6arzcqzsy2vng3k553xp",